+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "unruley" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [v0.0.4] - 2021-23-06
+- Fixed bug where existing ruler settings were being overwritten by the extension's default settings.
+- Added an alert when conflicting settings are found and allows the user to choose which settings to use with Unruley.
+- Fixed bug where the extension was overriding the default VS Code keybinding for thee command palette.
 
-- Initial release
+## [v0.0.2] - 2021-10-06
+- Initial release]